requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
idna==2.8
oauthlib==2.1.0
pep8==1.7.1
PySocks==1.6.8
python-dotenv==0.10.1
requests==2.21.0
requests-oauthlib==1.0.0
rope==0.11.0
six==1.12.0
slackclient==1.3.0
tweepy==3.7.0
urllib3==1.24.1
websocket-client==0.54.0